# server/main.js
"use strict";

let express         = require('express');
let bodyparser      = require('body-parser');
let path            = require('path');
let mongoose        = require('mongoose');

let logger          = require('./src/Utils/Logger/logger');
let validator       = require('./src/Utils/Validators/index');

let storeService            = require('./src/Services/store/index');
let userService             = require('./src/Services/user/index');
let productService          = require('./src/Services/product/index');
let encouragementService    = require('./src/Services/encouragements/index');
let messageService          = require('./src/Services/messages/index');
let shiftSerice             = require('./src/Services/shift/index');

let app = express();

app.set('view engine', 'ejs');
let port = 3000;
app.use(bodyparser.json());
app.listen(port);
app.locals.baseurl = "http://localhost:" + port;
app.locals.mongourl = 'mongodb://localhost/IBBLS';



_connectToDb();
_setapApiEndpoints();

logger.info('server is now running on port: ', {'port': port});


function _connectToDb(){
    mongoose.Promise = global.Promise;
    mongoose.connect(app.locals.mongourl);
    let db = mongoose.connection;
    db.on('error', function(){
        console.error.bind(console, 'connection error:');
        throw 'Cant connect to DB...';
    });
    db.once('open', function() {
        console.log('connected to db successfuly');
    });
}

function _setapApiEndpoints() {

    app.use('/scripts', express.static(path.join(__dirname, '/../scripts')));

    app.get('/', function (req, res) {
        res.sendFile(path.join(__dirname, '../index.html'));
    });

//User Services
    app.post('/user/login', async function (req, res) {
        if (!validator.login(req.body)) {
            res.status(404).send('invalid parameters');
            return;
        }
        let result = await userService.login(req.body.username, req.body.password);
        if(result.sessionId != null){
            res.status(200).send(result);
        }
        else{
            res.status(401).send();
        }
    });

    app.post('/user/logout', async function (req, res) {
        if (!validator.sessionId(req.body)){
            res.status(404).send('invalid parameters');
            return;
        }
        let result = await userService.logout(req.body.sessionId);
        if(result.code == 200)
            res.status(200).send('logout succeeded');
        else
            res.status(result.code).send(result.err);
    });

    app.post('/user/retrievePassword', async function (req, res) {
        if (!validator.sessionId(req.body)) {
            res.status(404).send('invalid parameters');
            return;
        }
        let result = await userService.retrievePassword(req.body.sessionId);
        if(result.code == 200)
            res.status(200).send('retrieved password successfully');
        else
            res.status(result.code).send(result.err);
    });

    app.post('/user/changePassword', async function (req, res) {
        if (!validator.changePassword(req.body)){
            res.status(404).send('invalid parameters');
            return;
        }
        let result = await userService.changePassword(req.body.sessionId, req.body.oldPass, req.body.newPass);
        if(result.code == 200)
            res.status(200).send('changed password succeeded');
        else
            res.status(result.code).send(result.err);
    });

    app.get('/user/getProfile', async function (req, res) {
        if(!('sessionid' in req.headers)) {
            res.status(404).send('invalid parameters');
            return;
        }
        let result = await userService.getProfile(req.headers.sessionid);
        if(result.code == 200)
            res.status(200).send(result.user);
        else
            res.status(result.code).send(result.err);
    });

//Salesman Services
    app.post('/salesman/startShift', async function (req, res) {
        if(!validator.startOrEndShift(req.body)) {
            res.status(404).send('invalid parameters');
            return;
        }
        let result = await shiftSerice.startShift(req.body.sessionID, req.body.shift);
        if(result.code == 200)
            res.status(200).send();
        else
            res.status(result.code).send(result.err);
    });

    app.post('/salesman/finishShift', async function (req, res) {
        if(!validator.startOrEndShift(req.body)) {
            res.status(404).send('invalid parameters');
            return;
        }
        let result = await shiftSerice.endShift(req.body.sessionID, req.body.shift);
        if(result.code == 200)
            res.status(200).send();
        else
            res.status(result.code).send(result.err);
    });

    app.post('/salesman/reportSale', async function (req, res) {
        if(!validator.reportSaleOrOpened(req.body)) {
            res.status(404).send('invalid parameters');
            return;
        }
        let result = await shiftSerice.reportSale(req.body.sessionID, req.body.shiftId, req.body.productId, req.body.quantity);
        if(result.code == 200)
            res.status(200).send();
        else
            res.status(result.code).send(result.err);
    });

    app.post('/salesman/reportOpened', async function (req, res) {
        if(!validator.reportSaleOrOpened(req.body)) {
            res.status(404).send('invalid parameters');
            return;
        }
        let result = await shiftSerice.reportOpened(req.body.sessionID, req.body.shiftId, req.body.productId, req.body.quantity);
        if(result.code == 200)
            res.status(200).send();
        else
            res.status(result.code).send(result.err);
    });

    app.post('/salesman/addShiftComment', async function (req, res) {
        if(!validator.addShiftComment(req.body)) {
            res.status(404).send('invalid parameters');
            return;
        }
        let result = await shiftSerice.addShiftComment(req.body.sessionID, req.body.shiftId, req.body.content);
        if(result.code == 200)
            res.status(200).send();
        else
            res.status(result.code).send(result.err);
    });

    app.get('/salesman/:shiftId/activeShiftEncouragements', async function (req, res) {
        if(!('sessionid' in req.headers)) {
            res.status(404).send('invalid parameters');
            return;
        }
        let result = await shiftSerice.getActiveShiftEncouragements(req.body.sessionID, req.params.shiftId);
        if(result.code == 200)
            res.status(200).send(result.encouragements);
        else
            res.status(result.code).send(result.err);
    });

    app.get('/salesman/encouragements', function (req, res) {
        res.status(200).send('get encouragements list');
    });

    app.get('/salesman/getAllShifts', function (req, res) {
        res.status(200).send('get shifts list');
    });

    app.get('/salesman/getCurrentShift', async function (req, res) {
        if(!('sessionid' in req.headers)) {
            res.status(404).send('invalid parameters');
            return;
        }
        let result = await shiftSerice.getSalesmanCurrentShift(req.body.sessionID);
        if(result.code == 200)
            res.status(200).send(result.shift);
        else
            res.status(result.code).send(result.err);
    });

    app.post('/salesman/addShiftsConstraints', function (req, res) {
        res.status(200).send('add shifts constraints');
    });

    app.get('/salesman/salesHistory', function (req, res) {
        res.status(200).send('get sales history');
    });

    app.get('/salesman/getBroadcastMessages', async function (req, res) {
        if(!('sessionid' in req.headers)) {
            res.status(404).send('invalid parameters');
            return;
        }
        let sessionId = req.headers.sessionid;
        let result = await messageService.getInbox(sessionId);
        if(result.code == 200)
            res.status(200).send(result.inbox);
        else {
            res.status(result.code).send(result.err);
            messageService.markAsRead(sessionId);
        }
    });

    app.post('/salesman/shiftRegister', function (req, res) {
        res.status(200).send('registration to shift');
    });

//Management Services

    app.post('/management/addUser', async function (req, res) {
        if (!validator.addUser(req.body)) {
            res.status(404).send('invalid parameters');
            return;
        }
        let result = await userService.addUser(req.body.sessionId, req.body.userDetails);
        if(result.code == 200){
            res.status(200).send(result.user);
        }
        else{
            res.status(result.code).send(result.err);
        }
    });

    app.post('/management/editUser', async function (req, res) {
        if (!validator.editUser(req.body)) {
            res.status(404).send('invalid parameters');
            return;
        }
        let result = await userService.editUser(req.body.sessionId, req.body.username, req.body.userDetails);
        if(result.code == 200){
            res.status(200).send();
        }
        else{
            res.status(result.code).send(result.err);
        }
    });

    app.post('/management/deleteUser', async function (req, res) {
        if (!validator.deleteUser(req.body)) {
            res.status(404).send('invalid parameters');
            return;
        }
        let result = await userService.deleteUser(req.body.sessionId, req.body.username);
        if(result.code == 200){
            res.status(200).send();
        }
        else{
            res.status(result.code).send(result.err);
        }
    });

    app.get('/management/getAllUsers', async function(req, res){
        if(!('sessionid' in req.headers)) {
            res.status(404).send('invalid parameters');
            return;
        }
        let result = await userService.getAllUsers(req.headers.sessionid);
        if(result.code == 200){
            res.status(200).send(result.users);
        }
        else{
            res.status(result.code).send(result.err);
        }
    });

    app.post('/management/addStore', async function (req, res) {
        if (!validator.addOrEditOrDeleteStore(req.body)) {
            res.status(404).send('invalid parameters');
            return;
        }
        let result = await storeService.addStore(req.body.sessionId, req.body.storeDetails);
        if (result.err != null)
        {
            res.status(result.code).send(result.err);
        }
        else
        {
            res.status(result.code).send(result.store);
        }
    });

    app.post('/management/editStore', async function (req, res) {
        if (!validator.addOrEditOrDeleteStore(req.body)) {
            res.status(404).send('invalid parameters');
            return;
        }
        let result = await storeService.editStore(req.body.sessionId, req.body.storeDetails);
        if (result.err != null)
        {
            res.status(result.code).send(result.err);
        }
        else
        {
            res.status(result.code).send(result.store);
        }
    });

    app.post('/management/deleteStore', async function (req, res) {
        if (!validator.deleteStore(req.body)) {
            res.status(404).send('invalid parameters');
            return;
        }
        let result = await storeService.deleteStroe(req.body.sessionId, req.body.storeId);
        if (result.err != null)
        {
            res.status(result.code).send(result.err);
        }
        else
        {
            res.status(result.code).send(result.store);
        }
    });

    app.get('/management/getAllStores', async function (req, res) {
        if(!('sessionid' in req.headers)) {
            res.status(404).send('invalid parameters');
            return;
        }
        let result = await storeService.getAllStores(req.headers.sessionid);
        if (result.err != null) {
            res.status(result.code).send(result.err);
        }
        else {
            res.status(result.code).send(result.stores);
        }
    });

    app.get('/management/:storeId/getStore', async function (req, res) {
        if(!('sessionid' in req.headers)) {
            res.status(404).send('invalid parameters');
            return;
        }
        let result = await storeService.getStore(req.headers.sessionid, req.params.storeId);
        if (result.err != null) {
            res.status(result.code).send(result.err);
        }
        else {
            res.status(result.code).send(result.store);
        }
    });

    app.post('/management/addProduct', async function (req, res) {
        if (!validator.addOrEditProduct(req.body)) {
            res.status(404).send('invalid parameters');
            return;
        }
        let result = await productService.addProduct(req.body.sessionId, req.body.productDetails);
        if (result.err != null) {
            res.status(result.code).send(result.err);
        }
        else {
            res.status(result.code).send(result.product);
        }
    });

    app.post('/management/editProduct', async function (req, res) {
        if (!validator.addOrEditProduct(req.body)) {
            res.status(404).send('invalid parameters');
            return;
        }
        let result = await productService.editProduct(req.body.sessionId, req.body.productDetails);
        if (result.err != null) {
            res.status(result.code).send(result.err);
        }
        else {
            res.status(result.code).send(result.product);
        }
    });

    app.post('/management/deleteProduct', async function (req, res) {
        if (!validator.deleteProduct(req.body)) {
            res.status(404).send('invalid parameters');
            return;
        }
        let result = await productService.deleteProduct(req.body.sessionId, req.body.productId);
        if (result.err != null) {
            res.status(result.code).send(result.err);
        }
        else {
            res.status(result.code).send(result.product);
        }
    });

    app.get('/management/getAllProducts', async function (req, res) {
        if(!('sessionid' in req.headers)) {
            res.status(404).send('invalid parameters');
            return;
        }
        let result = await productService.getAllProducts(req.headers.sessionid);
        if (result.err != null) {
            res.status(result.code).send(result.err);
        }
        else {
            res.status(result.code).send(result.products);
        }
    });

    app.get('/management/:productId/getProduct', async function (req, res) {
        if(!('sessionid' in req.headers)) {
            res.status(404).send('invalid parameters');
            return;
        }
        let result = await productService.getProduct(req.headers.sessionid, req.params.productId);
        if (result.err != null) {
            res.status(result.code).send(result.err);
        }
        else {
            res.status(result.code).send(result.product);
        }
    });


    app.post('/management/addEncouragement', async function (req, res) {
        if (!validator.addOrEditEncouragement(req.body)) {
            res.status(404).send('invalid parameters');
            return;
        }
        let result = await encouragementServices.addEncouragement(req.body.sessionId, req.body.encouragementDetails);
        if (result.err != null) {
            res.status(result.code).send(result.err);
        }
        else {
            res.status(result.code).send(result.encouragement);
        }
    });

    app.post('/management/editEncouragement', async function (req, res) {
        if (!validator.addOrEditEncouragement(req.body)) {
            res.status(404).send('invalid parameters');
            return;
        }
        let result = await encouragementServices.editEncouragement(req.body.sessionId ,req.body.encouragementDetails);
        if (result.err != null) {
            res.status(result.code).send(result.err);
        }
        else {
            res.status(result.code).send(result.encouragement);
        }
    });

    app.post('/management/deleteEncouragement', async function (req, res) {
        if (!validator.deleteEncouragement(req.body)) {
            res.status(404).send('invalid parameters');
            return;
        }
        let result = await encouragementServices.deleteEncouragement(req.body.sessionId ,req.body.encouragementDetails);
        if (result.err != null) {
            res.status(result.code).send(result.err);
        }
        else {
            res.status(result.code).send(result.encouragement);
        }
    });

    app.get('/management/getAllEncouragements', async function (req, res) {
        if(!('sessionid' in req.headers)) {
            res.status(404).send('invalid parameters');
            return;
        }
        let result = await encouragementServices.getAllEncouragements(req.headers.sessionid);
        if (result.err != null) {
            res.status(result.code).send(result.err);
        }
        else {
            res.status(result.code).send(result.encouragements);
        }
    });

    app.get('/management/:encouragementId/getAllEncouragements', async function (req, res) {
        if(!('sessionid' in req.headers)) {
            res.status(404).send('invalid parameters');
            return;
        }
        let result = await encouragementServices.getEncouragement(req.headers.sessionid, req.params.encouragementId);
        if (result.err != null) {
            res.status(result.code).send(result.err);
        }
        else {
            res.status(result.code).send(result.encouragement);
        }
    });

    app.post('/management/addShifts', async function (req, res) {
        if (!validator.addOrPublishShifts(req.body)) {
            res.status(404).send('invalid parameters');
            return;
        }
        let result = await shiftSerice.addShifts(req.body.sessionId, req.body.shiftArr);
        if(result.code == 200)
            res.status(200).send(result.shiftArr);
        else
            res.status(result.code).send(result.err);
    });

    app.post('/management/publishShifts', async function (req, res) {
        if (!validator.addOrPublishShifts(req.body)) {
            res.status(404).send('invalid parameters');
            return;
        }
        let result = await shiftSerice.publishShifts(req.body.sessionId, req.body.shiftArr);
        if(result.code == 200)
            res.status(200).send();
        else
            res.status(result.code).send(result.err);
    });

    app.post('/management/editShifts', function (req, res) {
        res.status(200).send('edit user');
    });

    app.post('/management/deleteShift', function (req, res) {
        res.status(200).send('delete user');
    });

//Manager Services

    app.post('/manager/addNotificationRule', function (req, res) {
        res.status(200).send('add new notification rule');
    });

    app.post('/manager/removeNotificationRule', function (req, res) {
        res.status(200).send('remove notification rule');
    });

    app.post('/manager/setNotificationRule', function (req, res) {
        res.status(200).send('set new notification rule');
    });

    app.post('/manager/sendBroadcastMessage', async function (req, res) {
        if (!validator.sendBroadcastMessage(req.body)) {
            res.status(404).send('invalid parameters');
            return;
        }
        let result = await messageService.sendBroadcast(req.body.sessionId, req.body.content, req.body.date);
        if(result.code == 200)
            res.status(200).send();
        else
            res.status(result.code).send(result.err);
    });

    app.get('/manager/getShiftNotes', function (req, res) {
        res.status(200).send('get shift notes');
    });

    app.post('/manager/editSalesReport', function (req, res) {
        res.status(200).send('edit sales report');
    });

    app.get('/manager/getRecommendations', function (req, res) {
        res.status(200).send('TODO - not yet implemented');
    });

    app.get('/manager/getShiftDetails', function (req, res) {
        res.status(200).send('get a specific shift details');
    });

    app.get('/manager/getShortages', function (req, res) {
        res.status(200).send('get shortages');
    });

    app.post('/manager/publishShifts', function (req, res) {
        res.status(200).send('publish shifts');
    });

    app.get('/manager/getReports', function (req, res) {
        res.status(200).send('get reports of some kind');
    });
}